convert UICollectionViewDelegateFlowLayout to Swift
